Move event flow to its own singleton
